+++ src/Symfony/Component/Security/CHANGELOG.md
@@ -4,7 +4,8 @@ CHANGELOG
 3.3.0
 -----
 
- * deprecated `setVoters()` of the `AccessDecisionManager` class in favor of passing the voters to the constructor.
+ * deprecated `AccessDecisionManager::setVoters()` in favor of passing the
+   voters to the constructor.
 
 3.2.0
 -----